// generated/graphql.ts
export enum Ride_Step_Enum {
  RideEnded = 'RIDE_ENDED',
  RidePaused = 'RIDE_PAUSED',
  RideResumed = 'RIDE_RESUMED',
  RideStarted = 'RIDE_STARTED'
}
// ...
export type FetchAllHubsQuery = { __typename?: 'query_root', hubs: Array<{ __typename?: 'hubs', id: any, latitude?: any | null, longitude?: any | null, name?: string | null, organization_id: any, created_at: any }> };

export type FetchCompletedRidesQueryVariables = Exact<{
  id?: InputMaybe<Scalars['uuid']['input']>;
}>;


export type FetchCompletedRidesQuery = { __typename?: 'query_root', ride_details: Array<{ __typename?: 'ride_details', created_at: any, end_time?: any | null, end_hub_id?: any | null, id: any, scooter_id: any, start_hub_id: any, start_time: any, total_cost?: any | null, ride_steps: Array<{ __typename?: 'ride_steps', id: any, steps: Ride_Step_Enum }>, hubByStartHubId: { __typename?: 'hubs', latitude?: any | null, longitude?: any | null, name?: string | null, id: any } }> };